Fixed SOA query ID not written to wire and debug messages.

+++ libknot/zone/zone-contents.c
@@ -1766,7 +1766,7 @@ DEBUG_KNOT_ZONE(
 DEBUG_KNOT_ZONE(
 	//char *n = knot_dname_to_str(name_copy);
 	debug_knot_zone("Finding closest encloser..\nStarting with: %.*s\n", 
-	                name_size, name_tmp);
+			(int)name_size, name_tmp);
 	//free(n);
 );
 
@@ -1776,7 +1776,7 @@ DEBUG_KNOT_ZONE(
 DEBUG_KNOT_ZONE(
 		//char *n = knot_dname_to_str(name_copy);
 		debug_knot_zone("Chopped leftmost label: %.*s\n",
-		                  name_size, name_tmp);
+				(int)name_size, name_tmp);
 		//free(n);
 );
 		// not satisfied in root zone!!
@@ -2019,7 +2019,7 @@ int knot_zone_contents_tree_apply_postorder(knot_zone_contents_t *zone,
 /*----------------------------------------------------------------------------*/
 
 int knot_zone_contents_tree_apply_inorder(knot_zone_contents_t *zone,
-                              void (*function)(knot_node_t *node, void *data),
+			      void (*function)(knot_node_t *node, void *data),
                               void *data)
 {
 	if (zone == NULL) {
@@ -2073,7 +2073,7 @@ int knot_zone_contents_nsec3_apply_postorder(knot_zone_contents_t *zone,
 /*----------------------------------------------------------------------------*/
 
 int knot_zone_contents_nsec3_apply_inorder(knot_zone_contents_t *zone,
-                              void (*function)(knot_node_t *node, void *data),
+			      void (*function)(knot_node_t *node, void *data),
                               void *data)
 {
 	if (zone == NULL) {
